Extracting StewardsFileParser out of StewardsFile#parse
